build 1.0.2
Code refactoring
