+++ lib/Heap.rb
@@ -2,6 +2,7 @@ require 'Heap/version'
 
 # Main ruby-heap rb file
 module Heap
+  require 'Heap/heap_template'
   require 'Heap/binary_heap/binary_heap'
   require 'Heap/binary_heap/binary_heap_min'
   require 'Heap/binary_heap/binary_heap_max'
