+++ lib/Heap/binary_heap/binary_heap_min.rb
@@ -3,22 +3,11 @@ module Heap
     # Binary Heap with min root
     class MinHeap < BinaryHeap
       def extract_min
-        @elements[0]
+        extract_root
       end
 
       def extract_min!
-        swap(1, count)
-        el = @elements.pop
-        swim_down(1)
-        el
-      end
-
-      def sort
-        el_temp = @elements.clone
-        result = []
-        result.push extract_min! while count > 0
-        @elements = el_temp
-        result
+        extract_root!
       end
 
       private
