+++ lib/Heap/version.rb
@@ -1,3 +1,3 @@
 module Heap
-  VERSION = '1.0.1'
+  VERSION = '1.0.2'
 end
